Updated Updater and Reader formats
Updated the Strategy Pattern to use Updater classes instead of an 'Update' class. The updater classes no longer return a ShippingUpdate, but rather build one and then apply it toward the shipment if it exists.
Also separated the line reading/processing logic from the simulator into a separate Reader class.